afxdp/ha: relocate inline tests to ha_tests.rs
Same playbook. 322-LOC inline test block extracted to sibling ha_tests.rs.

After: ha.rs drops from 774 → 455 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
